Add PCB rev 2 pinout, fix "power down".
The line to disable the ADC didn't appear to actually do anything.
Approx. 60 µA would flow out of the CE and CS pins.

After removing the power-on LED and the LDO from the Arduino Pro Mini,
fixing the ADC, and setting some output pins to tristate (input), the
remaining current measured 150 nA, which is well below the autodischarge
rate of any typical battery and negligible compared to the consumption
of protection and charger circuits.